+++ lib/statistics/tasks.rb
@@ -2,3 +2,5 @@ module Statistics
   module Tasks
   end
 end
+
+require_relative 'tasks/connpass'
